pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "The SA/Swiss Chair on Blockchain Technology at the University of Johannesburg is a unique entity within the university, dedicated to advancing research and education in the field of blockchain with aim for real-world blockchain applications and development."
authors = ["UJ Blockchain <blockchain@uj.ac.za>"]
license = "Attribution-NonCommercial-ShareAlike 4.0"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.17"
bx-django-utils = "^81"
django-autoslug = "^1.9.9"
django-axes = "^7.0.0"
django-cleanup = "^9.0.0"
django-cors-headers = "^4.6.0"
django-import-export = "^4.2.1"
django-jazzmin = "^3.0.1"
django-recaptcha = "^4.0.0"
django-reversion = "^5.1.0"
django-reversion-compare = "^0.18.1"
django-csp = "^3.8"
gevent = "^24.11.1"
greenlet = "^3.1.1"
huey = "^2.5.2"
jinja2 = "^3.1.4"
pillow = "^11.0.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
python-decouple = "^3.8"
redis = "^5.2.0"
sentry-sdk = { extras = ["django"], version = "^2.18.0" }
shortuuid = "^1.0.13"
twilio = "^9.3.6"
validators = "^0.34.0"
dotmap = "^1.3.30"
django-split-settings = "^1.3.2"
django-auto-logout = "^0.5.1"
django-prose-editor = { extras = ["sanitize"], version = "^0.10.2" }
django-simple-history = "^3.7.0"
django-multiselectfield = "^0.1.13"
pre-commit = "^4.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
coverage = "^7.6.4"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
werkzeug = "^3.1.3"
model-bakery = "^1.20.0"


[tool.isort]
multi_line_output = 5
line_length = 119


[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
